+++ hybris/platforms/common/windowbuffer.h
@@ -23,6 +23,8 @@
 #ifndef WINDOWBUFFER_H
 #define WINDOWBUFFER_H
 
+#include <android-config.h>
+
 #include <cutils/native_handle.h>
 #include <string.h>
 #include <system/window.h>
@@ -36,7 +38,7 @@ class RemoteWindowBuffer : public BaseNativeWindowBuffer
 				unsigned int height,
 				unsigned int stride,
 				unsigned int format,
-				unsigned int usage,
+				uint64_t usage,
 				buffer_handle_t handle
 				) {
 			// Base members
